requirements.txt
# Base ----------------------------------------
torch>=1.8.1
torchvision>=0.9.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
gdown
flask
Werkzeug
argparse
transformers>=4.46.2  # for Mllama
ultralytics

# # Logging -------------------------------------
# tensorboard>=2.4.1
# comet-ml

# # Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

thop  # FLOPs computation
